refactor: coyote 패키지 정리
