+++ tomcat/src/main/java/org/apache/coyote/util/HttpResponseParser.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.util;
 
 import java.nio.charset.StandardCharsets;
 import java.util.StringJoiner;
